Port AudioSystem to current Ecs structure
