+++ Ecs/Audio/AudioSystem.cpp
@@ -1,0 +1,4 @@
+#include <NovelRT/Ecs/Audio/AudioEmitterState.hpp>
+#include <NovelRT/Ecs/Audio/Components/AudioEmitterStateComponent.hpp>
+#include <NovelRT/Ecs/Audio/Components/AudioEmitterComponent.hpp>
+#include <NovelRT/Ecs/Audio/AudioSystem.hpp>
