--- LegacySrc/include/NovelRT/Ecs/Audio/AudioEmitterStateComponent.h
@@ -1,42 +1,0 @@
-// Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root
-// for more information.
-
-#ifndef NOVELRT_ECS_AUDIO_AUDIOEMITTERSTATECOMPONENT_H
-#define NOVELRT_ECS_AUDIO_AUDIOEMITTERSTATECOMPONENT_H
-
-#ifndef NOVELRT_ECS_AUDIO_H
-#error NovelRT does not support including types explicitly by default. Please include Ecs.Audio.h instead for the Ecs.Audio namespace subset.
-#endif
-
-namespace NovelRT::Ecs::Audio
-{
-    struct AudioEmitterStateComponent
-    {
-        AudioEmitterState state = AudioEmitterState::Stopped;
-        float fadeDuration = 0;
-        float fadeExpectedVolume = 0;
-
-        inline AudioEmitterStateComponent& operator+=(const AudioEmitterStateComponent& other) noexcept
-        {
-            state = other.state;
-            fadeDuration = other.fadeDuration;
-            fadeExpectedVolume = other.fadeExpectedVolume;
-            return *this;
-        };
-
-        friend inline bool operator==(const AudioEmitterStateComponent& lhs,
-                                      const AudioEmitterStateComponent& rhs) noexcept
-        {
-            return lhs.state == rhs.state && lhs.fadeDuration == rhs.fadeDuration &&
-                   lhs.fadeExpectedVolume == rhs.fadeExpectedVolume;
-        }
-
-        friend inline bool operator!=(const AudioEmitterStateComponent& lhs,
-                                      const AudioEmitterStateComponent& rhs) noexcept
-        {
-            return !(lhs == rhs);
-        }
-    };
-}
-
-#endif // NOVELRT_ECS_AUDIO_AUDIOEMITTERSTATECOMPONENT_H
